:truck: Move dashboard widgets to a shared location.
Even though they are mostly model specific, I feel they are unique enough to warrant their own shared namespace.
