+++ web/src/components/dashboards/SystemAdminDashboard.vue
@@ -59,7 +59,7 @@ import { ref } from "vue"
 
 import userGroupsApi from "@/api/user-groups-api"
 
-import DatasetSearchCard from "@/components/datasets/DatasetSearchCard.vue"
+import DatasetSearchCard from "@/components/dashboard-widgets/DatasetSearchCard.vue"
 import QaScenariosCard from "@/components/qa-scenarios/QaScenariosCard.vue"
 
 const isLoadingUserGroups = ref(false)
